requirements.txt
groq==0.33.0
python-dotenv==1.0.1
requests==2.31.0